Remove CIRCTStage
